Move float16 into fluid folder (#8394)
* move float16 into fluid

* fix include

* move to platform folder
