+++ paddle/fluid/platform/CMakeLists.txt
@@ -27,7 +27,7 @@ ELSE()
     set(MKLDNN_CTX_DEPS)
 ENDIF()
 
-# memcpy deoends on device_context, here add deps individually for
+# memcpy depends on device_context, here add deps individually for
 # avoiding cycle dependencies
 cc_library(device_context SRCS device_context.cc DEPS memory buddy_allocator
     system_allocator memory_block meta_data meta_cache place eigen3 ${GPU_CTX_DEPS} ${MKLDNN_CTX_DEPS})
@@ -39,3 +39,6 @@ nv_test(nccl_test SRCS nccl_test.cu DEPS dynload_cuda gpu_info device_context)
 
 cc_library(profiler SRCS profiler.cc DEPS device_context)
 cc_test(profiler_test SRCS profiler_test.cc DEPS profiler)
+
+nv_test(float16_gpu_test SRCS float16_test.cu)
+cc_test(float16_test SRCS float16_test.cc)
