+++ paddle/fluid/platform/float16_test.cc
@@ -9,7 +9,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
 
-#include "paddle/math/float16.h"
+#include "paddle/fluid/platform/float16.h"
 
 #include <gtest/gtest.h>
